+++ darts/models/__init__.py
@@ -11,9 +11,10 @@ from darts.utils.utils import NotImportedModule
 
 try:
     # `lightgbm` needs to be imported first to avoid segmentation fault
-    from darts.models.forecasting.lgbm import LightGBMModel
+    from darts.models.forecasting.lgbm import LightGBMClassifierModel, LightGBMModel
 except ModuleNotFoundError:
     LightGBMModel = NotImportedModule(module_name="LightGBM", warn=False)
+    LightGBMClassifierModel = NotImportedModule(module_name="LightGBM", warn=False)
 
 # Forecasting
 from darts.models.forecasting.arima import ARIMA
@@ -35,7 +36,11 @@ from darts.models.forecasting.kalman_forecaster import KalmanForecaster
 from darts.models.forecasting.linear_regression_model import LinearRegressionModel
 from darts.models.forecasting.random_forest import RandomForest, RandomForestModel
 from darts.models.forecasting.regression_ensemble_model import RegressionEnsembleModel
-from darts.models.forecasting.sklearn_model import RegressionModel, SKLearnModel
+from darts.models.forecasting.sklearn_model import (
+    RegressionModel,
+    SKLearnClassifierModel,
+    SKLearnModel,
+)
 from darts.models.forecasting.theta import FourTheta, Theta
 from darts.models.forecasting.varima import VARIMA
 
@@ -83,9 +88,14 @@ except ImportError:
     Prophet = NotImportedModule(module_name="Prophet", warn=False)
 
 try:
-    from darts.models.forecasting.catboost_model import CatBoostModel
+    from darts.models.forecasting.catboost_model import (
+        CatBoostClassifierModel,
+        CatBoostModel,
+    )
 except ModuleNotFoundError:
     CatBoostModel = NotImportedModule(module_name="CatBoost", warn=False)
+    CatBoostClassifierModel = NotImportedModule(module_name="CatBoost", warn=False)
+
 
 try:
     from darts.models.forecasting.sf_auto_arima import AutoARIMA
@@ -116,9 +126,10 @@ except ImportError:
     AutoTBATS = NotImportedModule(module_name="StatsForecast", warn=False)
 
 try:
-    from darts.models.forecasting.xgboost import XGBModel
+    from darts.models.forecasting.xgboost import XGBClassifierModel, XGBModel
 except ImportError:
     XGBModel = NotImportedModule(module_name="XGBoost")
+    XGBClassifierModel = NotImportedModule(module_name="XGBoost")
 
 # Filtering
 from darts.models.filtering.gaussian_process_filter import GaussianProcessFilter
@@ -127,6 +138,7 @@ from darts.models.filtering.moving_average_filter import MovingAverageFilter
 
 __all__ = [
     "LightGBMModel",
+    "LightGBMClassifierModel",
     "ARIMA",
     "NaiveDrift",
     "NaiveMean",
@@ -140,6 +152,7 @@ __all__ = [
     "RandomForestModel",
     "RegressionEnsembleModel",
     "SKLearnModel",
+    "SKLearnClassifierModel",
     "RegressionModel",
     "TBATS",
     "FourTheta",
@@ -161,6 +174,7 @@ __all__ = [
     "TSMixerModel",
     "Prophet",
     "CatBoostModel",
+    "CatBoostClassifierModel",
     "Croston",
     "AutoARIMA",
     "AutoCES",
@@ -170,6 +184,7 @@ __all__ = [
     "AutoTBATS",
     "StatsForecastModel",
     "XGBModel",
+    "XGBClassifierModel",
     "GaussianProcessFilter",
     "KalmanFilter",
     "MovingAverageFilter",